Feature/account page

+++ pages/success.jsx
@@ -1,32 +1,63 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import Link from "next/link";
 import { BsBagCheckFill } from "react-icons/bs";
 import { useRouter } from "next/router";
 import { useStateContext } from "../context/StateContext";
 import { runFireworks } from "../lib/fireworks";
+import { useSession } from "next-auth/react";
 
 const Success = () => {
   const router = useRouter();
   const { setCartItems, cartItems } = useStateContext();
+  const { data: session } = useSession({ required: true });
 
   const processData = async () => {
     if (!router.isReady) return;
     if (!router.query?.session_id) return;
 
-    const response = await fetch(`/api/stripe?key=${router.query.session_id}`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
+    // get stripe checkout information
+    const stripeResponse = await fetch(
+      `/api/stripe/orders?key=${router.query.session_id}`,
+      {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
+    if (stripeResponse.statusCode === 500) return;
+    const stripeData = await stripeResponse.json();
 
-    if (response.statusCode === 500) return;
-    const data = await response.json();
-
-    if (data.session.payment_status === "paid") {
+    if (stripeData.session.payment_status === "paid") {
       runFireworks();
 
       if (cartItems.length) {
+        // get information about the logged in user
+        const response = await fetch(
+          `/api/prisma/user?key=${session.user.email}`,
+          {
+            method: "GET",
+            headers: { "Content-Type": "application/json" },
+          }
+        );
+        if (response.statusCode === 500) return;
+        const prismaUserData = await response.json();
+
+        // store the data in prisma
+        await Promise.all(
+          cartItems.map(async (product) => {
+            await fetch(`/api/prisma/order/`, {
+              method: "POST",
+              headers: { "Content-Type": "application/json" },
+              body: JSON.stringify({
+                cartItems: product,
+                stripeData,
+                userData: prismaUserData,
+              }),
+            });
+          })
+        );
+
         // update the stock amount in sanity
         const sanityCheck = await fetch("/api/sanityUpdate", {
           method: "POST",
@@ -35,7 +66,10 @@ const Success = () => {
           },
           body: JSON.stringify(cartItems),
         });
+        if (sanityCheck.statusCode === 500) return;
+        await sanityCheck.json();
 
+        // clear out cart items
         setCartItems([]);
         localStorage.clear();
       }
@@ -44,7 +78,7 @@ const Success = () => {
 
   useEffect(() => {
     processData();
-  }, [router.isReady]);
+  }, [session]); // ! this may not be necessary when actually deployed
 
   return (
     <div className="success-wrapper">
